refactor(broker): extract shared session lifecycle into broker/session.rs (#556)
Extract duplicated session lifecycle logic from UDS (broker/mod.rs) and WS
(broker/ws/mod.rs) into a shared broker/session.rs module. Both transports
now delegate to session_setup(), process_inbound_message(),
process_oneshot_send(), and session_teardown().

This fixes the WS path missing subscribe_mentioned() before broadcast —
a bug identified in the #552 audit.

Includes 20 tests: 15 auto-subscribe tests (moved from mod.rs) plus 5 new
tests for session_setup, session_teardown, process_inbound_message, and
process_oneshot_send.
